chore: brand with Debcraft
- Replace instances of Starcraft with Debcraft.
- Introduce a minimal README.md
