--- .github/CODEOWNERS
@@ -1,9 +1,0 @@
-# GitHub CODEOWNERS. See:
-# https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners
-# Remember, the last match takes precedence.
-
-# Documentation owners
-/docs/  @medubelko @canonical/starcraft
-
-# Finally, all CODEOWNERS changes need to be approved by The Man, Himself.
-/.github/CODEOWNERS     @sergiusens
