+++ .github/workflows/tics.yaml
@@ -2,14 +2,13 @@ name: TICS
 on:
   push:
     branches:
-      - "main"
+      # - "main"
       # For troubleshooting
       - "work/tiobe**"
 
 jobs:
   TICS:
     uses: canonical/starflow/.github/workflows/tics.yaml@main
-    # Uncomment this and add your project's name on Tiobe TICS
-    # with:
-    #   project: "starcraft"
-    secrets: inherit
+    with:
+      project: "debcraft"
+      secrets: inherit
