--- README-template.md
@@ -1,62 +1,0 @@
-<If the software has approved imagery or logo, add it here.>
-
-# Starcraft
-
-<Include the build status badges.>
-
-<Answer elevator-pitch questions about the software – What is it? What does it do? What
-essential problem does it solve? Who is it useful for?>
-
-<If it's a library or framework and the solutions it provides are more open-ended, you
-can describe the project's overarching philosophy.>
-
-## Basic usage
-
-<Show the essence of the product with carefully-chosen examples of it in use. Give the
-prospective user a sense of what the software _feels_ like.>
-
-<This could be the constituent parts of the main user story, or smaller individual
-stories. Some examples:>
-
-- <For CLI apps and tools, three-to-five invocations of essential commands>
-- <For frameworks, some short examples of the structures and syntax>
-- <For libraries, the library name (`import <package>`) and the main APIs>
-
-## Installation
-
-<Briefly describe what platforms the software is supported on.>
-
-<Provide the simplest way to install it, prioritising the snap release on Ubuntu.>
-
-<If available, link to other installation methods in the documentation.>
-
-## Documentation
-
-<Encourage the reader to start their journey with the documentation and link to it.>
-
-## Community and support
-
-<Invite the user to participate in the community, with a link to the Code of Conduct and
-its central forums.>
-
-You can report any issues or bugs on the project's [GitHub
-repository](https://github.com/canonical/starcraft/issues).
-
-Starcraft is covered by the [Ubuntu Code of
-Conduct](https://ubuntu.com/community/ethos/code-of-conduct).
-
-## Contribute to Starcraft
-
-Starcraft is open source and part of the Canonical family. We would love your help.
-
-If you're interested, start with the [contribution guide](HACKING.md).
-
-We welcome any suggestions and help with the docs. The [Canonical Open Documentation
-Academy](https://github.com/canonical/open-documentation-academy) is the hub for doc
-development, including Starcraft docs. No prior coding experience is required.
-
-## License and copyright
-
-Starcraft is released under the [GPL-3.0 license](LICENSE).
-
-© <start year>-<current year> Canonical Ltd.
